refactor: move tokens to separate file and turn display implementation into feature
