+++ examples/inner-content/Cargo.toml
@@ -3,10 +3,8 @@ name = "inner-content"
 version = "0.1.0"
 edition = "2021"
 
-
 [dependencies]
-ytml-lang={path = "../../packages/ytml-lang"}
 
 [[bin]]
-name = "inner_content"
+name = "inner-content"
 path = "main.rs"
