+++ examples/read-file/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-ytml-lang = {path = "../../packages/ytml-lang", features = ["fs"]}
 
 [[bin]]
 name = "read-file"
