+++ packages/ytml-cli/src/main.rs
@@ -72,4 +72,3 @@ fn main() -> notify::Result<()> {
         }
     }
 }
-
