+++ packages/ytml-lang/Cargo.toml
@@ -9,3 +9,4 @@ pest_derive = "2.7.2"
 
 [features]
 fs = []
+display = []
