+++ packages/ytml-lang/src/html/mod.rs
@@ -1,4 +1,4 @@
-use super::{Tag, TagInnerElement};
+use crate::tokens::{Tag, TagInnerElement};
 
 pub fn ast_to_html(ast: Vec<Tag>, indent: usize) -> String {
     let mut html_content = String::new();
